refactor: move CarManager from Controller to Domain
