+++ src/main/java/racingcar/domain/CarManager.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.domain;
 
 import racingcar.domain.Car;
 import racingcar.domain.CarMovement;
